+++ src/main/java/com/revature/shoply/customer/DTOs/IncomingCartItemDTO.java
@@ -1,4 +1,4 @@
-package com.revature.shoply.cart.DTOs;
+package com.revature.shoply.customer.DTOs;
 
 import java.util.UUID;
