+++ src/main/java/com/revature/shoply/product/service/ProductTagService.java
@@ -1,8 +1,8 @@
 package com.revature.shoply.product.service;
 
 import com.revature.shoply.product.repository.ProductDAO;
-import com.revature.shoply.Repositories.ProductTagDAO;
-import com.revature.shoply.Repositories.TagDAO;
+import com.revature.shoply.repositories.ProductTagDAO;
+import com.revature.shoply.repositories.TagDAO;
 import com.revature.shoply.models.Product;
 import com.revature.shoply.models.ProductTag;
 import com.revature.shoply.models.Tag;
