+++ src/main/java/com/revature/shoply/repositories/OrderDAO.java
@@ -1,4 +1,4 @@
-package com.revature.shoply.Repositories;
+package com.revature.shoply.repositories;
 
 import java.util.UUID;
